requirements.txt
numpy==1.15.1
Pillow==7.0.0
